Project/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
backcall=0.1.0=py37_0
bzip2=1.0.8=h516909a_1
ca-certificates=2019.11.27=0
cairo=1.16.0=hfb77d84_1002
certifi=2019.11.28=py37_0
cffi=1.13.1=py37h2e261b9_0
chardet=3.0.4=pypi_0
cudatoolkit=10.0.130=0
dbus=1.13.6=he372182_0
decorator=4.4.1=py_0
django=2.2.5=py37_1
expat=2.2.5=he1b5a44_1004
facenet-pytorch=0.4.0=pypi_0
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.0=he983fc9_1
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=py37h6f030ca_1002
gmp=6.1.2=hf484d3e_1000
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=hf484d3e_1000
gst-plugins-base=1.14.5=h0935bb2_0
gstreamer=1.14.5=h36ae1b5_0
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.5=nompi_h3c11f04_1104
icu=64.2=he1b5a44_1
idna=2.8=pypi_0
intel-openmp=2019.4=243
ipython=7.9.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.15.1=py37_0
jpeg=9c=h14c3975_1001
lame=3.100=h14c3975_1001
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libffi=3.2.1=he1b5a44_1006
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_2
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_openblas
liblapacke=3.8.0=14_openblas
libopenblas=0.3.7=h6e990d7_3
libpng=1.6.37=hed695b0_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=hfc65ed5_0
libuuid=2.32.1=h14c3975_1000
libwebp=1.0.2=h576950b_1
libxcb=1.13=h14c3975_1002
libxml2=2.9.10=hee79883_0
lz4-c=1.8.3=he1b5a44_1001
mkl=2019.4=243
ncurses=6.1=hf484d3e_1002
nettle=3.4.1=h1bed415_1002
ninja=1.9.0=py37hfd86e86_0
numpy=1.17.3=py37h95a1406_0
olefile=0.46=py37_0
opencv=4.1.0=py37h3aa1047_6
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1d=h7b6447c_3
pandas=0.25.3=py37he6710b0_0
parso=0.5.1=py_0
pcre=8.43=he1b5a44_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.2.1=py37h34e0f95_0
pip=19.3.1=py37_0
pixman=0.38.0=h516909a_1003
prompt_toolkit=2.0.10=py_0
pthread-stubs=0.4=h14c3975_1001
ptyprocess=0.6.0=py37_0
pycparser=2.19=py37_0
pygments=2.4.2=py_0
python=3.7.3=h33d41f4_1
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.7_cuda10.0.130_cudnn7.6.3_0
pytz=2019.3=py_0
qt=5.9.7=h0c104cb_3
readline=8.0=hf8c457e_0
requests=2.22.0=pypi_0
setuptools=41.6.0=py37_1
six=1.12.0=py37_0
sqlite=3.30.1=hcee41ef_0
sqlparse=0.3.0=py_0
tk=8.6.9=hed695b0_1003
torchvision=0.4.2=py37_cu100
tqdm=4.38.0=py_0
traitlets=4.3.3=py37_0
urllib3=1.25.6=pypi_0
wcwidth=0.1.7=py37_0
wheel=0.33.6=py37_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_1001
zlib=1.2.11=h516909a_1006
zstd=1.4.3=h3b9ef0a_0
